@@ -1,22 +1,22 @@
 module github.com/cybozu/neco-containers/bmc-reverse-proxy
 
-go 1.17
+go 1.19
 
 require (
-	github.com/cybozu-go/log v1.6.1
-	github.com/cybozu-go/well v1.11.0
-	k8s.io/apimachinery v0.24.8
-	k8s.io/client-go v0.24.8
+	github.com/cybozu-go/log v1.7.0
+	github.com/cybozu-go/well v1.11.2
+	k8s.io/apimachinery v0.25.6
+	k8s.io/client-go v0.25.6
 )
 
 require (
 	github.com/PuerkitoBio/purell v1.1.1 // indirect
 	github.com/PuerkitoBio/urlesc v0.0.0-20170810143723-de5bf2ad4578 // indirect
-	github.com/cybozu-go/netutil v1.4.2 // indirect
+	github.com/cybozu-go/netutil v1.4.4 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/emicklei/go-restful v2.9.5+incompatible // indirect
-	github.com/fsnotify/fsnotify v1.5.1 // indirect
-	github.com/go-logr/logr v1.2.0 // indirect
+	github.com/emicklei/go-restful/v3 v3.8.0 // indirect
+	github.com/fsnotify/fsnotify v1.6.0 // indirect
+	github.com/go-logr/logr v1.2.3 // indirect
 	github.com/go-openapi/jsonpointer v0.19.5 // indirect
 	github.com/go-openapi/jsonreference v0.19.5 // indirect
 	github.com/go-openapi/swag v0.19.14 // indirect
